+++ plate-editor/plugin/align-plugin.ts
