+++ plate-editor/plugin/fixed-toolbar-plugin.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { FixedToolbar } from "@/plate-editor/ui/fixed-toolbar";
+import { FixedToolbarButtons } from "@/plate-editor/ui/fixed-toolbar-buttons";
+import { createPlatePlugin } from "@udecode/plate/react";
+
+export const FixedToolbarPlugin = createPlatePlugin({
+    key: "fixed-toolbar",
+    render: {
+        beforeEditable: () => (
+            <FixedToolbar>
+                <FixedToolbarButtons />
+            </FixedToolbar>
+        ),
+    },
+});
