+++ plate-editor/plugin/floating-toolbar-plugin.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { FloatingToolbar } from "@/plate-editor/ui/floating-toolbar";
+import { FloatingToolbarButtons } from "@/plate-editor/ui/floating-toolbar-buttons";
+import { createPlatePlugin } from "@udecode/plate/react";
+
+export const FloatingToolbarPlugin = createPlatePlugin({
+    key: "floating-toolbar",
+    render: {
+        afterEditable: () => (
+            <FloatingToolbar>
+                <FloatingToolbarButtons />
+            </FloatingToolbar>
+        ),
+    },
+});
