+++ plate-editor/plugin/reset-block-type-plugin.ts
