+++ plate-editor/plugin/soft-break-plugin.ts
