+++ plate-editor/plugin/toc-plugin.ts
@@ -1,0 +1,10 @@
+"use client";
+
+import { TocPlugin } from "@udecode/plate-heading/react";
+
+export const tocPlugin = TocPlugin.configure({
+  options: {
+    // isScroll: true,
+    topOffset: 80,
+  },
+});
