--- plugin/fixed-toolbar-plugin.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { FixedToolbar } from "@/components/ui/fixed-toolbar";
-import { FixedToolbarButtons } from "@/components/ui/fixed-toolbar-buttons";
-import { createPlatePlugin } from "@udecode/plate/react";
-
-export const FixedToolbarPlugin = createPlatePlugin({
-  key: "fixed-toolbar",
-  render: {
-    beforeEditable: () => (
-      <FixedToolbar>
-        <FixedToolbarButtons />
-      </FixedToolbar>
-    ),
-  },
-});
