--- plugin/floating-toolbar-plugin.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { FloatingToolbar } from "@/components/ui/floating-toolbar";
-import { FloatingToolbarButtons } from "@/components/ui/floating-toolbar-buttons";
-import { createPlatePlugin } from "@udecode/plate/react";
-
-export const FloatingToolbarPlugin = createPlatePlugin({
-  key: "floating-toolbar",
-  render: {
-    afterEditable: () => (
-      <FloatingToolbar>
-        <FloatingToolbarButtons />
-      </FloatingToolbar>
-    ),
-  },
-});
